// src/interFace/interFace.ts
interface Submenu {
  title: string;
  link: string;
  previewImg?: StaticImageData;
  pluseIncon?: boolean;
  megaMenu?: Submenu[];
}

// Define the structure for each menu item
export interface MenuItem {
  id: number;
  hasDropdown: boolean;
  children: boolean;
  active: boolean;
  title: string;
  pluseIncon: boolean;
  link: string;
  columnFour?: boolean;
  previewImg?: boolean;
  submenus?: Submenu[];
  megaMenu?: boolean;
  pageLayout?: boolean;
  lastDropdown?: boolean;
}